requirements.txt
python-telegram-bot==20.7
pandas==2.1.4
requests==2.31.0
google-generativeai==0.3.2